Enable static generation of home/[lang]'s Page component
